+++ Tests/LinuxMain.swift
@@ -360,6 +360,7 @@ extension RulesTests {
         ("testBlockBasedKVO", testBlockBasedKVO),
         ("testClassDelegateProtocol", testClassDelegateProtocol),
         ("testClosingBrace", testClosingBrace),
+        ("testClosureBodyLength", testClosureBodyLength),
         ("testClosureEndIndentation", testClosureEndIndentation),
         ("testClosureParameterPosition", testClosureParameterPosition),
         ("testClosureSpacing", testClosureSpacing),